Finished swarming/hypersearch separation
