+++ src/nupic/swarming/hypersearch/object_json.py
