+++ src/nupic/swarming/utils.py
